refactor: Change test directory structure (#25)
