+++ swumeal/src/test/java/com/swumeal/app/domain/menu/service/MenuServiceTests.java
@@ -1,7 +1,6 @@
-package com.swumeal.app.service;
+package com.swumeal.app.domain.menu.service;
 
 import com.swumeal.app.domain.menu.dto.MenuDTO;
-import com.swumeal.app.domain.menu.service.MenuService;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
