+++ swumeal/src/test/java/com/swumeal/app/global/mapper/MenuMapperTests.java
@@ -1,4 +1,4 @@
-package com.swumeal.app.mapper;
+package com.swumeal.app.global.mapper;
 
 import com.swumeal.app.domain.menu.domain.MenuItemVO;
 import com.swumeal.app.domain.menu.domain.MenuVO;
@@ -6,7 +6,6 @@ import com.swumeal.app.domain.menu.dto.MenuDTO;
 import com.swumeal.app.domain.menu.vo.MenuDataVo;
 import com.swumeal.app.domain.model.MealTypeEnum;
 import com.swumeal.app.domain.model.TimeEnum;
-import com.swumeal.app.global.mapper.MenuMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
